Fix(Use cases): Add authorization validation for use cases

+++ src/domain/course-management/application/use-cases/delete-course.ts
@@ -1,4 +1,5 @@
 import { left, right, type Either } from '@/core/either'
+import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found-error'
 import { type UseCase } from '@/core/use-cases/use-case'
 import { type Course } from '../../enterprise/entities/course'
@@ -6,10 +7,11 @@ import { type CoursesRepository } from '../repositories/courses-repository'
 
 interface DeleteCourseUseCaseRequest {
   courseId: string
+  instructorId: string
 }
 
 type DeleteCourseUseCaseResponse = Either<
-ResourceNotFoundError,
+ResourceNotFoundError | NotAllowedError,
 {
   course: Course
 }
@@ -21,12 +23,19 @@ export class DeleteCourseUseCase implements UseCase<DeleteCourseUseCaseRequest, 
   ) { }
 
   async exec({
-    courseId
+    courseId,
+    instructorId
   }: DeleteCourseUseCaseRequest): Promise<DeleteCourseUseCaseResponse> {
     const course = await this.coursesRepository.findById(courseId)
 
     if (!course) {
       return left(new ResourceNotFoundError())
+    }
+
+    const instructorIsTheOwner = course.instructorId.toString() === instructorId
+
+    if (!instructorIsTheOwner) {
+      return left(new NotAllowedError())
     }
 
     await this.coursesRepository.delete(course)
